backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Redis
redis==5.0.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 任务调度
apscheduler==3.10.4

# 数据模拟
faker==22.0.0

# 测试
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Excel处理
openpyxl==3.1.2

# 工具
python-dotenv==1.0.0
